remove AbstractInvokeStrategy
resolves https://github.com/prooph/service-bus/issues/152
