--- src/Plugin/InvokeStrategy/AbstractInvokeStrategy.php
@@ -1,41 +1,0 @@
-<?php
-/**
- * This file is part of the prooph/service-bus.
- * (c) 2014-2017 prooph software GmbH <contact@prooph.de>
- * (c) 2015-2017 Sascha-Oliver Prolic <saschaprolic@googlemail.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Prooph\ServiceBus\Plugin\InvokeStrategy;
-
-use Prooph\Common\Event\ActionEvent;
-use Prooph\ServiceBus\MessageBus;
-use Prooph\ServiceBus\Plugin\AbstractPlugin;
-
-abstract class AbstractInvokeStrategy extends AbstractPlugin
-{
-    /**
-     * @param mixed $handler
-     * @param mixed $message
-     */
-    abstract protected function invoke($handler, $message): void;
-
-    public function attachToMessageBus(MessageBus $messageBus): void
-    {
-        $this->listenerHandlers[] = $messageBus->attach(
-            MessageBus::EVENT_DISPATCH,
-            function (ActionEvent $actionEvent): void {
-                $message = $actionEvent->getParam(MessageBus::EVENT_PARAM_MESSAGE);
-                $handler = $actionEvent->getParam(MessageBus::EVENT_PARAM_MESSAGE_HANDLER);
-
-                $this->invoke($handler, $message);
-                $actionEvent->setParam(MessageBus::EVENT_PARAM_MESSAGE_HANDLED, true);
-            },
-            MessageBus::PRIORITY_INVOKE_HANDLER
-        );
-    }
-}
